+++ env.sh
@@ -1,0 +1,2 @@
+export PATH=$(realpath ./pico/picotool/install/bin):$PATH
+export PICO_SDK_PATH=$(realpath ./pico/pico-sdk)