+++ test/control/test-callcc-1.scm
@@ -1,0 +1,4 @@
+(write
+  (+ (* 4 4)
+     (call/cc (lambda (k) (* 3 3)))))
+(newline)