+++ test/control/test-callcc-3.scm
@@ -1,0 +1,7 @@
+(write
+  (call/cc
+    (lambda (k)
+      (k 0)
+      (write #f)
+      (newline))))
+(newline)