+++ test/control/test-callcc-4.scm
@@ -1,0 +1,3 @@
+(let ((f (lambda (k) 0)))
+  (write (call/cc f))
+  (newline))