+++ test/control/test-callcc-5.scm
@@ -1,0 +1,3 @@
+(let ((f (lambda (k) (k 0) (write #f) (newline))))
+  (write (call/cc f))
+  (newline))