+++ test/control/test-callcc-values-0.scm
@@ -1,0 +1,4 @@
+(call-with-values
+  (lambda () (call/cc (lambda (k) (k))))
+  (lambda x (write x)))
+(newline)