+++ test/values/test-values-1.scm
@@ -1,0 +1,4 @@
+(let ((writeln (lambda (x) (write x) (newline))))
+  (call-with-values
+    (lambda () (values 2 3 5))
+    (lambda (x y z) (writeln x) (writeln y) (writeln z) 0)))