+++ test/values/test-values-unwoke.scm
@@ -1,0 +1,5 @@
+(let ()
+  (call-with-values
+    (lambda () 0)
+    (lambda (x) (write x) (newline)))
+)